functions/requirements.txt
# Flask and Werkzeug compatibility
flask==2.2.5
werkzeug==2.2.3
markupsafe==2.1.3

# CORS support
flask-cors==4.0.0

# Environment variables
python-dotenv==1.0.0

# Your AI/ML dependencies
langgraph>=0.0.40
langchain-google-genai>=1.0.0

# Firebase dependencies
firebase-functions>=0.1.0
firebase-admin>=6.2.0

# Optional dependencies
grandalf>=0.8
rich>=13.0.0